+++ source/Messaging.Infrastructure/OutgoingMessages/Requesting/SendSuccessNotification.cs
@@ -16,41 +16,37 @@ using System;
 using System.Text.Json.Serialization;
 using Messaging.Application.Common.Commands;
 
-namespace Messaging.Infrastructure.OutgoingMessages;
+namespace Messaging.Infrastructure.OutgoingMessages.Requesting;
 
-public class SendFailureNotification : InternalCommand
+public class SendSuccessNotification : InternalCommand
 {
     [JsonConstructor]
-    public SendFailureNotification(Guid id, Guid requestId, string idempotencyId, string failureDescription, string reason, string referenceId, string messageType)
+    public SendSuccessNotification(Guid id, Guid requestId, string idempotencyId, string referenceId, string messageType, Uri messageStorageLocation)
         : base(id)
     {
         RequestId = requestId;
         IdempotencyId = idempotencyId;
-        FailureDescription = failureDescription;
-        Reason = reason;
         ReferenceId = referenceId;
         MessageType = messageType;
+        MessageStorageLocation = messageStorageLocation;
     }
 
-    public SendFailureNotification(Guid requestId, string idempotencyId, string failureDescription, string reason, string referenceId, string messageType)
+    public SendSuccessNotification(Guid requestId, string idempotencyId, string referenceId, string messageType, Uri messageStorageLocation)
     {
         RequestId = requestId;
         IdempotencyId = idempotencyId;
-        FailureDescription = failureDescription;
-        Reason = reason;
         ReferenceId = referenceId;
         MessageType = messageType;
+        MessageStorageLocation = messageStorageLocation;
     }
 
     public Guid RequestId { get; }
 
     public string IdempotencyId { get; }
 
-    public string FailureDescription { get; }
-
-    public string Reason { get; }
-
     public string ReferenceId { get; }
 
     public string MessageType { get; }
+
+    public Uri MessageStorageLocation { get; }
 }
